requirements.txt
Flask>=3.0,<4
requests>=2.31,<3
node-semver>=0.7.0,<1
